+++ lib/src/style/layer/symbol_layer.dart
@@ -649,12 +649,10 @@ class SymbolLayer extends Layer {
 
   /// Specifies an uniform elevation from the ground, in meters.
   /// Default value: 0. Minimum value: 0.
-  @experimental
   double? symbolZOffset;
 
   /// Specifies an uniform elevation from the ground, in meters.
   /// Default value: 0. Minimum value: 0.
-  @experimental
   List<Object>? symbolZOffsetExpression;
 
   /// The color with which the text will be drawn.